Use the same time range as in the WHERE clause for segment lookup
If a query has a time range in the WHERE clause, only a small subset of the
segments will be queried. The coverage implementation, however, will still
try to guarantee that all segments have been covered, causing a Insufficient
segment coverage error.

In order to avoid this, the query time range is being used for segment lookup.